day 15: keep a count of goblins and elves
The loop can now exit early when a unit tries to do something, but there
are no more opponents on the board.  That way, `rounds` is exactly the
number of rounds that fully completed, since it's only incremented at
the end of the loop.

All the example tests pass after this.