day 17: be less generous when switching to Resting
If left and right of the current position are all
water-water-water-(water-)*-clay, then we know that the current position
should be Resting.

However, this is now *under*-counting, because after flooding a row, it
doesn't re-visit cells and set them to Resting, so when they're
dripped-on from above, the row above does not flood left and right.